onTime only applies to live stream
Notes from email:

live stream has 'onTime' setting in the config that makes it stop after
300 seconds (5 mins)

When playing the replay clips, we want to remove this setting as it is
causing the clips to stop playing if you jump past 5 mins in the player
controls when viewing a video.

So can you please see if we can have this onTime setting included when
the live surfcam pages loads, but then remove it when a replay clip is
clicked?

https://support.jwplayer.com/customer/portal/questions/9481763-changing-
configuration-options-after-setup-

https://stackoverflow.com/questions/26425908/jwplayer-change-configurati
on-dynamically-without-reloading-video-skin-caption

might need to destroy the player re-init