Updates and bug fixes
Changes requested:

* On the live page, the first thumbnail doesn't seem to do anything
when clicked, so we can't play the first replay of the day, unless you
click another thumbnail and then click back on the first one.

* On the live page, when you click a replay thumbnail, can you please
set the 'replays' button to active, and set the 'live' button to
non-active.

* With the replays slider, I know that I asked you to hide it when the
page loads if there is no data available, but I'd like you update this
so that it still displays the date and time dropdowns. This is because
sometimes there won't be data for the current day (eg. if the camera is
down), but then we still want the user to be able to view the clips
from the previous days.

	- Also, on the replays page, if there is no initial data for the
current day then this means there won't be a clip available for when
the page first 	loads. Is it possible to just show the player but have
nothing loaded in it?

* In the mornings (early, eg. 5am or 6am when there isn't any clips
available yet), it looks like the js is doing 2 requests to get the
data - the first request is for the current day and then I'm assuming
(because the data is empty) that it then tries to pull back data for
the previous day? This is resulting in the time dropdown showing the
list of times form the previous day, but the date dropdown showing
today's date. Does that make sense? Can you please check this and fix
it?

----

Ben wants to add Google Analytics tracking to each click event on a
thumbnail, which is pretty straight forward in jquery, but not sure
about Vue.js. I had a quick look and there is some info on the web, and
there would be a lot more similar articles I think:
https://forum.vuejs.org/t/best-way-to-use-google-analytics-event-trackin
g-in-vue-js/20523/2

According to the GA docs (
https://developers.google.com/analytics/devguides/collection/analyticsjs
/events ) I think we want to use the 'non-interaction events', so
something along the lines of...

ga('send', 'event', 'Replays', 'Surfcam replay thumbnail clicked',
'Snapper Rocks - 2018-02-05 04:52:24',{
  nonInteraction: true
});

Where 'Surfcam replay thumbnail clicked' is the eventAction, and the
'Snapper Rocks - 2018-02-05 04:52:24' is the eventLabel, which pulls
together the surfcam name, and the date/time of the thumbnail.